requirements.txt
Flask==2.1.1
boto3==1.18.65
python-dotenv==0.19.1
elasticsearch==7.14.0